got the correct command to run
